scripts: Remove obsolete build scripts
This commit removes the Zephyr SDK build scripts that are obsolete and
no longer used with the new GitHub Actions-based CI.

In general, one should refrain from building the Zephyr SDK locally
because the build environment configurations are very complex to set up
and there is really no point in building the SDK locally when there is
a CI infrastructure to do exactly that.

If it is ever necessary to build the Zephyr SDK locally for some
reason, one should use a local GitHub Actions runner, such as
nektos/act.

Signed-off-by: Stephanos Ioannidis <root@stephanos.io>