[WIP] ci: Upload release assets in parts
This commit modifies the CI release workflow to upload the release
assets in multiple parts because the `softprops/action-gh-release`
action attempts to load all specified release assets into the runner
memory at once and this may cause the runner instance to run out of
memory.

For more details, refer to the GitHub issue #520.

Revert this when the action is updated to use the streams.

Signed-off-by: Stephanos Ioannidis <root@stephanos.io>